Use arrays instead of vectors for custom animation data
The "current_canim_indices" and "current_canim_values" variables only
store a maximum of 99999 elements, which isn't enough even for the one
custom animation that currently comes with Saturn.

For now, we'll use vectors instead of arrays to store both the custom
animation's indices and values. However, it may be possible to entirely
rewrite the way the animations are loaded in the future.

Co-authored-by: bjrkk <cfxqpkjq@gmail.com>