saturn_bind_texture: Return pointers to literals directly instead of constructing string classes from them
The saturn_bind_texture function constructed a string from the contents
of a literal and then returned a pointer to its data if the texture to
replace is either one of the playable character's textures or the
background of the Secret Aquarium. Doing this isn't a good idea since
the string's destructor will get called once it goes out of scope,
freeing the data it returns.

The pointer to any of the string literals really should've been
returned instead from the start, but since that wasn't the case, this
commit fixes that problem.